requirements.txt
matplotlib==3.10.8
numpy==2.4.0
pandas==2.3.3
Pillow==12.0.0
plotly==6.3.1
streamlit==1.52.1
wordcloud==1.9.4
